+++ tooling/lib/build.ts
@@ -11,6 +11,8 @@ export async function build(root: string): Promise<void> {
       join(root, 'app/bin/leia.ts'),
       join(root, 'app/lib/app.ts'),
       join(root, 'app/lib/compiler.ts'),
+      join(root, 'app/lib/runtime.ts'),
+      join(root, 'app/lib/api.ts'),
     ],
     root: join(root, 'app'),
     outdir,